Make sure U-Boot and U-Boot fw tools agree on environment size.
This issue is somewhat complicated. The heart of the problem is that
the size of the environment is board specific, and is only defined in
C code, which is very hard to get to from the bb recipe. So we go for
a different approach, we define the environment size in the recipe as
well and only verify that they agree in C code.

However, there are still complicating factors. In order for this check
to work, the size and file name has to be defined in a place where
both the u-boot recipe and mender-sdimg.bbclass can find it, which
means it can't be in either one, since they are both local. So I opted
for a different approach, moving both BOOTENV_SIZE and
IMAGE_BOOT_ENV_FILE to the mender-uboot.bbclass recipe, which is
global. That means I'm also trying to handle most of the details
(config file creation and environment creation) in the u-boot recipe,
leaving mender-sdimg to concentrate on image creation and less about
U-Boot details.

One consequence of creating the config file in the u-boot recipe, is
that U-boot is now a runtime package dependency, not just a deployment
dependency, hence the extra RDEPENDS. But this appears to be the
intention from the poky side all along, since it has code exactly for
dealing with this config file.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>