Pregenerate SSH keys for all QEMU images.
By doing it in the build instead of during image launch, we can
provide keys even for artifacts that will be installed in the future,
not just the currently launched image. This should provide a much
needed performance boost for QEMU testing.

The key generation on the QEMU device is particularly problematic
because of our rapid connection rate. Each new SSH connections spawns
another daemon, and even if the key generation eventually finishes, we
may end up with a lot of connections that waste CPU and can even cause
the instance to run out of memory. By having the keys up front, the
very first connection should succeed, eliminating the need for
additional connections.

Changelog: Title

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>