feat: Introduce `MENDER_EFI_LOADER`.
This is a partial restore of 197c998454bd818, but now guarded by the
`MENDER_EFI_LOADER` variable, so that configurations using edk2, or
something outside of Yocto, can also work.

Changelog: Introduce `MENDER_EFI_LOADER`. This can be used to specify
which bootloader is used to load other EFI apps, such as the grub-efi
bootloader. This is normally provided by firmware on x86, and is
therefore empty, but on ARM it is often set to u-boot or edk2.

Cancel-changelog: 197c998454bd8181a827228b39a3400f3e61ea1b

Ticket: None

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>